Minor changes to adapt to x86

+++ cpp/SampleIME/CandidateWindow.cpp
@@ -28,13 +28,13 @@ DWORD HighlightedCandidateColor(DWORD accentColor)
     int r = accentColor % 0x100;
     int g = (accentColor >> 8) % 0x100;
     int b = (accentColor >> 16) % 0x100;
-    float ratio = 0.75 * 510 / (min(r, min(g, b)) + max(r, max(g, b)));
+    float ratio = 0.75f * (float) (510 / (min(r, min(g, b)) + max(r, max(g, b))));
     BOOL maxxed = 0;
     while (ratio > 1 && !maxxed) {
         float fr = r * ratio;
         float fg = g * ratio;
         float fb = b * ratio;
-        float th = 255.999;
+        float th = 255.999f;
         float fm = max(fr, max(fg, fb));
         if (fm >= th) {
             float ft = fr + fg + fb;
@@ -48,7 +48,7 @@ DWORD HighlightedCandidateColor(DWORD accentColor)
             }
         }
         r = int(fr); g = int(fg); b = int(fb);
-        float newRatio = 0.75 * 510 / (min(r, min(g, b)) + max(r, max(g, b)));
+        float newRatio = 0.75f * (float) (510 / (min(r, min(g, b)) + max(r, max(g, b))));
         if (ratio == newRatio)
             maxxed = 1;
         else
